src/test/java/ArticleRestControllerTests.java
import cf.creativeflow.marquis.blog.MarquisApplication;
import cf.creativeflow.marquis.blog.controller.ArticleRestController;
import cf.creativeflow.marquis.blog.domain.Article;
import org.junit.Before;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.http.MediaType;
import org.springframework.http.converter.HttpMessageConverter;
import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
import org.springframework.test.context.web.WebAppConfiguration;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
import org.springframework.web.context.WebApplicationContext;

import java.nio.charset.Charset;
import java.util.ArrayList;
import java.util.List;

import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

@RunWith(SpringJUnit4ClassRunner.class)
@WebMvcTest(ArticleRestController.class)
@SpringBootTest(classes = MarquisApplication.class)
@WebAppConfiguration
public class ArticleRestControllerTests {

    private MediaType contentType = new MediaType(MediaType.APPLICATION_JSON.getType(),
            MediaType.APPLICATION_JSON.getSubtype(),
            Charset.forName("utf8"));

    @Autowired
    private MockMvc mockMvc;

    private HttpMessageConverter mappingJackson2HttpMessageConverter;

    private List<Article> articleListList = new ArrayList<>();

    //@Autowired
    //private ArticleRepository articleRepository;

    @Autowired
    private WebApplicationContext webApplicationContext;

    @Before
    public void setup() throws Exception {

    }

    @Test
    public void getAllArticles() throws Exception {

        mockMvc.perform(MockMvcRequestBuilders
                .get("/posts")
                .accept(contentType))
                .andExpect(status().isOk())
                .andExpect(MockMvcResultMatchers.jsonPath("$.posts").exists())
                .andExpect(MockMvcResultMatchers.jsonPath("$.posts[*].id").isNotEmpty());
    }


}
